+++ src/k-provider/ovp/response-types/kaltura-media-entry.js
@@ -125,6 +125,12 @@ export class KalturaMediaEntry {
   createdAt: number;
 
   /**
+   * @member - Entry updation date as Unix timestamp (In seconds)
+   * @type {number}
+   */
+  updatedAt: number;
+
+  /**
    * @member - Entry end date as Unix timestamp (In seconds)
    * @type {number}
    */
@@ -141,6 +147,18 @@ export class KalturaMediaEntry {
    * @type {number}
    */
   views: number;
+
+  /**
+   * @member - The download URL of the entry.
+   * @type {string}
+   */
+  downloadUrl: string;
+
+  /**
+   * @member - ID of entry creator.
+   * @type {string}
+   */
+  creatorId: string;
 
   /**
    * @constructor
@@ -162,8 +180,11 @@ export class KalturaMediaEntry {
     this.dvrStatus = entry.dvrStatus;
     this.tags = entry.tags;
     this.createdAt = entry.createdAt;
+    this.updatedAt = entry.updatedAt;
+    this.creatorId = entry.creatorId;
     this.endDate = entry.endDate;
     this.plays = entry.plays;
     this.views = entry.views;
+    this.downloadUrl = entry.downloadUrl;
   }
 }